requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
charset-normalizer==2.1.1
discord.py @ git+https://github.com/Rapptz/discord.py@e0341c9ef25c0093bff991ff09c87dba0414e58c
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
mysql-connector-python==8.0.30
numpy==1.23.3
pandas==1.4.4
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
yarl==1.8.1